+++ isitfit/cli/cost.py
@@ -30,7 +30,13 @@ def cost(ctx, filter_region):
 # Check note above about ndays
 # Another note about ndays: using click.IntRange for validation. Ref: https://click.palletsprojects.com/en/7.x/options/?highlight=prompt#range-options
 @cost.command(help='Analyze AWS EC2 cost', cls=IsitfitCommand)
-@click.option('--ndays', default=90, prompt='Number of days to lookback (use `isitfit cost analyze --ndays=90` to skip this prompt)', help='number of days to look back in the data history', type=click.IntRange(1, 90))
+@click.option(
+  '--ndays',
+  default=7,
+  prompt='Number of days to lookback (between 1 and 90, use `isitfit cost analyze --ndays=7` to skip this prompt)',
+  help='number of days to look back in the data history',
+  type=click.IntRange(1, 90)
+)
 @click.option('--filter-tags', default=None, help='filter instances for only those carrying this value in the tag name or value')
 @click.option('--save-details', is_flag=True, help='Save details behind calculations to CSV files')
 @click.pass_context
@@ -65,7 +71,13 @@ def analyze(ctx, ndays, filter_tags, save_details):
 
 # check note above about ndays
 @cost.command(help='Generate recommendations of optimal EC2 sizes', cls=IsitfitCommand)
-@click.option('--ndays', default=90, prompt='Number of days to lookback (use `isitfit cost optimize --ndays=90` to skip this prompt)', help='number of days to look back in the data history', type=click.IntRange(1, 90))
+@click.option(
+  '--ndays',
+  default=7,
+  prompt='Number of days to lookback (between 1 and 90, use `isitfit cost optimize --ndays=7` to skip this prompt)',
+  help='number of days to look back in the data history',
+  type=click.IntRange(1, 90)
+)
 @click.option('--n', default=-1, help='number of underused ec2 optimizations to find before stopping. Skip to get all optimizations')
 @click.option('--filter-tags', default=None, help='filter instances for only those carrying this value in the tag name or value')
 @click.pass_context